Refactor: Moves LogRowContext and Alert to grafana/ui
